transcription/requirements.txt
torch>=2.0.0
torchaudio>=2.0.0
nemo_toolkit[asr]
gradio>=4.0.0
